requirements.txt
beautifulsoup4==4.6.0
decorator==4.1.2
Django==1.11.5
django-filter==1.1.0
djangorestframework==3.7.3
feedparser==5.1.3
html2text==2017.10.4
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
Markdown==2.6.8
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.13.1
wcwidth==0.1.7
lxml==4.1.1
schedule==0.4.3
django-debug-toolbar

